pyproject.toml
[tool.black]
line-length = 88
include = 'src.*py$|tests.*py$'

[tool.poetry]
name = "car-azimuth-predictor-experimentation"
version = "0.1.0"
description = "Car azimuth estimation"
authors = ["Ivan Orlov <orlov.van@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8,<3.9"
tensorflow-macos = {markers = "sys_platform == \"darwin\" and platform_machine == \"arm64\"", version = ">=2.8.0"}
tensorflow = {version = ">=2.8.0", markers = "platform_machine != \"arm64\""}
albumentations = "1.2.1"
scipy = "1.8.1"
numpy = ">=1.23.1"
omegaconf = "^2.3.0"
pandas = "1.4.3"
tqdm = "^4.66.2"
tensorflow-probability = "0.16.0"
hydra-core = "1.2.0"
tensorflow-io-gcs-filesystem = "0.34"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
